Better devcontainer installs

+++ scripts/deps_ffmpeg.sh
@@ -2,12 +2,19 @@
 
 set -e
 
-rm -rf ffmpeg
-curl https://ffmpeg.org/releases/ffmpeg-7.0.tar.xz | tar xJ
-mv ffmpeg-7.0 ffmpeg
-pushd ffmpeg
-mkdir build
-./configure --prefix=${PWD}/build --pkg-config-flags="--static" --enable-debug --extra-cflags="-g" --enable-nonfree --enable-gpl --enable-libx264 --enable-libvpx --enable-libfdk-aac --disable-stripping --disable-decoder=exr,phm
-make -j$(nproc)
-make install
-popd
+if [ ! -f ffmpeg/build/bin/ffmpeg ]; then
+    rm -rf ffmpeg
+    curl https://ffmpeg.org/releases/ffmpeg-7.0.tar.xz | tar xJ
+    mv ffmpeg-7.0 ffmpeg
+    pushd ffmpeg
+    mkdir build
+    ./configure --prefix=${PWD}/build --pkg-config-flags="--static" --enable-debug --extra-cflags="-g" --enable-nonfree --enable-gpl --enable-libx264 --enable-libvpx --enable-libfdk-aac --disable-stripping --disable-decoder=exr,phm
+    make -j$(nproc)
+    make install
+    popd
+else
+    echo "ffmpeg already built, installing"
+    pushd ffmpeg
+    make install
+    popd
+fi
